answers/pyproject.toml
[tool.poetry]
name = "answers"
version = "0.1.0"
description = ""
authors = ["rychanya <rychanya@gmail.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.1"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
aiofiles = "^0.7.0"
celery = "^5.1.2"
redis = "^3.5.3"
certifi = "^2021.10.8"
pymongo = {extras = ["srv"], version = "^3.12.0"}
PyJWT = {extras = ["crypto"], version = "^2.3.0"}
python-multipart = "^0.0.5"
openpyxl = "^3.0.9"
python-dotenv = "^0.19.1"
Jinja2 = "^3.0.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^21.9b0", allow-prereleases = true}
pytest-mock = "^3.6.1"
isort = "^5.10.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
